Amend file parser signature

+++ include/file_parser.h
@@ -2,7 +2,7 @@
 #define AOC_FILE_PARSER_
 
 /* ~~~~~~~~~~~~~~~~ File parser ~~~~~~~~~~~~~~~ */
-typedef void(*AoC_parserFun)(char* line, int len, void* args);
+typedef void(*AoC_parserFun)(char* line, size_t len, void* args);
 int AoC_parseFile(char* file_name, AoC_parserFun parserFunction, void* args);
 
 #endif // AOC_FILE_PARSER_
